requirements.txt
PyYAML
google-cloud-storage
Flask

